refactor(embedding): improve logging and simplify voyage embedding implementation
