+++ crates/http-api-bindings/src/embedding/mod.rs
@@ -1,6 +1,5 @@
 mod llama;
 mod openai;
-mod voyage;
 
 use core::panic;
 use std::sync::Arc;
@@ -9,8 +8,8 @@ use llama::LlamaCppEngine;
 use tabby_common::config::HttpModelConfig;
 use tabby_inference::Embedding;
 
-use self::{openai::OpenAIEmbeddingEngine, voyage::VoyageEmbeddingEngine};
 use super::rate_limit;
+use openai::OpenAIEmbeddingEngine;
 
 pub async fn create(config: &HttpModelConfig) -> Arc<dyn Embedding> {
     let engine = match config.kind.as_str() {
@@ -30,16 +29,16 @@ pub async fn create(config: &HttpModelConfig) -> Arc<dyn Embedding> {
             config.api_key.as_deref(),
         ),
         "ollama/embedding" => ollama_api_bindings::create_embedding(config).await,
-        "voyage/embedding" => VoyageEmbeddingEngine::create(
-            config.api_endpoint.as_deref(),
+        "voyage/embedding" => OpenAIEmbeddingEngine::create(
+            config
+                .api_endpoint
+                .as_deref()
+                .unwrap_or("https://api.voyageai.com/v1"),
             config
                 .model_name
                 .as_deref()
                 .expect("model_name must be set for voyage/embedding"),
-            config
-                .api_key
-                .clone()
-                .expect("api_key must be set for voyage/embedding"),
+            config.api_key.as_deref(),
         ),
         unsupported_kind => panic!(
             "Unsupported kind for http embedding model: {}",
@@ -51,4 +50,11 @@ pub async fn create(config: &HttpModelConfig) -> Arc<dyn Embedding> {
         engine,
         config.rate_limit.request_per_minute,
     ))
+}
+
+#[macro_export]
+macro_rules! embedding_info_span {
+    ($kind:expr) => {
+        tracing::info_span!("embedding", kind = $kind)
+    };
 }
