+++ crates/http-api-bindings/src/embedding/openai.rs
@@ -1,14 +1,16 @@
 use anyhow::Context;
-use async_openai::{
-    config::OpenAIConfig,
-    types::{CreateEmbeddingRequest, EmbeddingInput},
-};
 use async_trait::async_trait;
+use reqwest::Client;
+use serde::{Deserialize, Serialize};
 use tabby_inference::Embedding;
 use tracing::{info_span, Instrument};
 
+use crate::embedding_info_span;
+
 pub struct OpenAIEmbeddingEngine {
-    client: async_openai::Client<OpenAIConfig>,
+    client: Client,
+    api_endpoint: String,
+    api_key: String,
     model_name: String,
 }
 
@@ -18,41 +20,69 @@ impl OpenAIEmbeddingEngine {
         model_name: &str,
         api_key: Option<&str>,
     ) -> Box<dyn Embedding> {
-        let config = OpenAIConfig::default()
-            .with_api_base(api_endpoint)
-            .with_api_key(api_key.unwrap_or_default());
-
-        let client = async_openai::Client::with_config(config);
-
+        let client = Client::new();
         Box::new(Self {
             client,
+            api_endpoint: format!("{}/embeddings", api_endpoint),
+            api_key: api_key.unwrap_or_default().to_owned(),
             model_name: model_name.to_owned(),
         })
     }
 }
 
+#[derive(Debug, Serialize)]
+struct EmbeddingRequest {
+    input: Vec<String>,
+    model: String,
+}
+
+#[derive(Debug, Deserialize)]
+struct EmbeddingResponse {
+    data: Vec<EmbeddingData>,
+}
+
+#[derive(Debug, Deserialize)]
+struct EmbeddingData {
+    embedding: Vec<f32>,
+}
+
 #[async_trait]
 impl Embedding for OpenAIEmbeddingEngine {
     async fn embed(&self, prompt: &str) -> anyhow::Result<Vec<f32>> {
-        let request = CreateEmbeddingRequest {
+        let request = EmbeddingRequest {
+            input: vec![prompt.to_owned()],
             model: self.model_name.clone(),
-            input: EmbeddingInput::String(prompt.to_owned()),
-            encoding_format: None,
-            user: None,
-            dimensions: None,
         };
-        let resp = self
+
+        let request_builder = self
             .client
-            .embeddings()
-            .create(request)
-            .instrument(info_span!("embedding", kind = "openai"))
+            .post(&self.api_endpoint)
+            .json(&request)
+            .header("content-type", "application/json")
+            .bearer_auth(&self.api_key);
+
+        let response = request_builder
+            .send()
+            .instrument(embedding_info_span!("openai"))
             .await?;
-        let data = resp
+
+        if !response.status().is_success() {
+            let status = response.status();
+            let error = response.text().await?;
+            return Err(anyhow::anyhow!("Error {}: {}", status.as_u16(), error));
+        }
+
+        let response_body = response
+            .json::<EmbeddingResponse>()
+            .await
+            .context("Failed to parse response body")?;
+
+        response_body
             .data
             .into_iter()
             .next()
-            .context("Failed to get embedding")?;
-        Ok(data.embedding)
+            .map(|data| data.embedding)
+            .ok_or_else(|| anyhow::anyhow!("No embedding data found"))
     }
 }
 
@@ -72,5 +102,14 @@ mod tests {
         );
         let embedding = engine.embed("Hello, world!").await.unwrap();
         assert_eq!(embedding.len(), 768);
+    }
+
+    #[tokio::test]
+    #[ignore]
+    async fn test_voyage_embedding() {
+        let api_key = std::env::var("VOYAGE_API_KEY").expect("VOYAGE_API_KEY must be set");
+        let engine = OpenAIEmbeddingEngine::create("https://api.voyageai.com/v1", "voyage-code-2", Some(&api_key));
+        let embedding = engine.embed("Hello, world!").await.unwrap();
+        assert_eq!(embedding.len(), 1536);
     }
 }
